Making some small cleanups and bug fixes.

+++ src/java/org/jnativehook/NativeHookException.java
@@ -29,55 +29,55 @@ public class NativeHookException extends Exception {
 	private static final long serialVersionUID = -7944506738553380757L;
 
 	/** The error code associated with this exception. */
-	private short code = NativeHookException.UNKNOWN_FAILURE;
+	private int code = NativeHookException.UNKNOWN_FAILURE;
 
 	/**
 	 * Unknown failure error code.
 	 *
 	 * @since	1.2
 	 */
-	public static final short UNKNOWN_FAILURE				= 0x00;
+	public static final int UNKNOWN_FAILURE					= 0x00;
 
 	/**
 	 * General failure error code.
 	 *
 	 * @since	1.2
 	 */
-	public static final short HOOK_FAILURE					= 0x01;
+	public static final int HOOK_FAILURE					= 0x01;
 
 	/**
 	 * X11 open display error code.
 	 *
 	 * @since	1.2
 	 */
-	public static final short X11_OPEN_DISPLAY				= 0x20;
+	public static final int X11_OPEN_DISPLAY				= 0x20;
 
 	/**
 	 * X11 extension XRecord was not found.
 	 *
 	 * @since	1.2
 	 */
-	public static final short X11_RECORD_NOT_FOUND			= 0x21;
+	public static final int X11_RECORD_NOT_FOUND			= 0x21;
 
 	/**
 	 * X11 extension XRecord could not allocate range.
 	 *
 	 * @since	1.2
 	 */
-	public static final short X11_RECORD_ALLOC_RANGE		= 0x22;
+	public static final int X11_RECORD_ALLOC_RANGE			= 0x22;
 
 	/**
 	 * X11 extension XRecord could not create a context.
 	 *
 	 * @since	1.2
 	 */
-	public static final short X11_RECORD_CREATE_CONTEXT		= 0x23;
+	public static final int X11_RECORD_CREATE_CONTEXT		= 0x23;
 
 
 	/**
 	 * X11 extension XRecord could not enable the context.
 	 */
-	public static final short X11_RECORD_ENABLE_CONTEXT		= 0x24;
+	public static final int X11_RECORD_ENABLE_CONTEXT		= 0x24;
 
 
 	/**
@@ -85,33 +85,33 @@ public class NativeHookException extends Exception {
 	 *
 	 * @since	1.2
 	 */
-	public static final short WIN_SET_HOOK					= 0x30;
+	public static final int WIN_SET_HOOK					= 0x30;
 
 
 	/**
 	 * Apple access for assistive devices is disabled.
 	 */
-	public static final short DARWIN_AXAPI_DISABLED			= 0x40;
+	public static final int DARWIN_AXAPI_DISABLED			= 0x40;
 
 	/**
 	 * Apple could not create an event port.
 	 */
-	public static final short DARWIN_CREATE_EVENT_PORT		= 0x41;
+	public static final int DARWIN_CREATE_EVENT_PORT		= 0x41;
 
 	/**
 	 * Apple could not create a run loop source.
 	 */
-	public static final short DARWIN_CREATE_RUN_LOOP_SOURCE	= 0x42;
+	public static final int DARWIN_CREATE_RUN_LOOP_SOURCE	= 0x42;
 
 	/**
 	 * Apple could not acquire the current run loop.
 	 */
-	public static final short DARWIN_GET_RUNLOOP			= 0x43;
+	public static final int DARWIN_GET_RUNLOOP				= 0x43;
 
 	/**
 	 * Apple could not create an observer.
 	 */
-	public static final short DARWIN_CREATE_OBSERVER		= 0x44;
+	public static final int DARWIN_CREATE_OBSERVER			= 0x44;
 
 
 	/**
@@ -130,7 +130,7 @@ public class NativeHookException extends Exception {
 	 *
 	 * @param code The native error code.
 	 */
-	public NativeHookException(short code) {
+	public NativeHookException(int code) {
 		super();
 
 		this.code = code;
@@ -154,7 +154,7 @@ public class NativeHookException extends Exception {
 	 * @param code The native error code.
 	 * @param message The detail message.
 	 */
-	public NativeHookException(short code, String message) {
+	public NativeHookException(int code, String message) {
 		super(message);
 
 		this.code = code;
@@ -179,7 +179,7 @@ public class NativeHookException extends Exception {
 	 * @param message The detail message.
 	 * @param cause The cause of the exception. A null value is permitted, and indicates that the cause is unknown.
 	 */
-	public NativeHookException(short code, String message, Throwable cause) {
+	public NativeHookException(int code, String message, Throwable cause) {
 		super(message, cause);
 
 		this.code = code;
@@ -199,7 +199,7 @@ public class NativeHookException extends Exception {
 	 *
 	 * @param cause The cause of the exception. A null value is permitted, and indicates that the cause is unknown.
 	 */
-	public NativeHookException(short code, Throwable cause) {
+	public NativeHookException(int code, Throwable cause) {
 		super(cause);
 
 		this.code = code;
@@ -212,7 +212,7 @@ public class NativeHookException extends Exception {
 	 *
 	 * @return The native error code
 	 */
-	public short getCode() {
+	public int getCode() {
 		return this.code;
 	}
 }
